Moved to dual.py remove duplication merge
